Add correction so lines are translated appropriately
This problem arises because before the Hough transformation, points
are shifted so that they are centered around the origin. The Hough
transformation accumulator then describes lines in this translated
system. The new Line constructor accepts a translation argument which
will create a line with the given parameters, translate it, and then
store the new parameters.