Update threshold behavior for Hough transform
Keep track of which points are already part of a line (judging by their
distance to the line). Require that the number of *newly found* points
on a line be greater than the threshold. The old behavior was that the
number of votes in the Hough accumulator be greather than the threshold.